week_2/day_5/Car_Dealership_project/pyproject.toml
[project]
name = "car-dealership-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
]